server/requirements.txt
stripe
Flask
flask-cors
python-dotenv
